LPS-52265 move packages
