+++ modules/portal/portal-security-ldap/src/com/liferay/portal/security/ldap/DefaultLDAPSettings.java
