+++ modules/portal/portal-security-ldap/src/com/liferay/portal/security/ldap/DefaultPortalLDAP.java
